Add user interface, url and view for adding related cards
This commit adds a button "add a related phrase" to the "card detail"
template. The button shows a modal, much like the "add a translation"
motal. The modal contains a form from which users can select a related
card, and submit to a new URL, adding this card to the original card's
`see_also` many-to-many relation.

This is the very beginning of this feature. It doesn't have a lot of
polish or any validation.