Add chosen.js and let users add multiple related cards at once
In this commit we add chosen.js and its required CSS. The input for
adding related cards is now a multiple select, and the corresponding
view takes multiple card IDs as inputs, makes sure those cards exist,
and returns messages with the number of successful card additions
and the number of errors.

There are still some UI imperfections (CSS perhaps? jquery version?),
the user feedback could be more accurate (like not double-counting
if, because of some edge case, a relation-addition was attempted
twice without error), the user's form should show existing relations
either as disabled options or in another section, and the "required
field" validation should show up (the native HTML validation doesn't
show up now that it's a chosen.js field).